chore: rename files and resolve conflicts
