+++ test/get_method_test.cpp
